requirements.txt
pandas
web3
requests
python-dotenv